[#1447] extracted presenter from create activity
