+++ app/src/main/java/org/akvo/flow/presentation/geoshape/entities/AreaShape.java
@@ -17,31 +17,31 @@
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.presentation.geoshape.create.entities;
+package org.akvo.flow.presentation.geoshape.entities;
 
 import android.os.Parcel;
 
 import java.util.List;
 
-public class PointShape extends Shape {
+public class AreaShape extends Shape {
 
-    public PointShape(String featureId, List<ShapePoint> points) {
+    public AreaShape(String featureId, List<ShapePoint> points) {
         super(featureId, points);
     }
 
-    public PointShape(Parcel in) {
+    public AreaShape(Parcel in) {
         super(in);
     }
 
-    public static final Creator<PointShape> CREATOR = new Creator<PointShape>() {
+    public static final Creator<AreaShape> CREATOR = new Creator<AreaShape>() {
         @Override
-        public PointShape createFromParcel(Parcel in) {
-            return new PointShape(in);
+        public AreaShape createFromParcel(Parcel in) {
+            return new AreaShape(in);
         }
 
         @Override
-        public PointShape[] newArray(int size) {
-            return new PointShape[size];
+        public AreaShape[] newArray(int size) {
+            return new AreaShape[size];
         }
     };
 }
